refactor(warnings): split tabular detectors
Move numeric-table and dot-leader warning rules into a focused tabular detector module. These rules share flattened table structure concerns, and the split keeps the central warning orchestrator smaller without changing warning output.
